Move fonts into src
