+++ .zeroignore
@@ -1,3 +1,4 @@
 utils
 components
+fonts
 server.js
